requirements.txt
# =============================================================
# agentic-rag — Python Dependencies
# LangChain + LangGraph + Pinecone
# =============================================================
# Install all dependencies:
#   pip install -r requirements.txt
# =============================================================


# -------------------------------------------------------------
# LangChain Core
# -------------------------------------------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# -------------------------------------------------------------
# LangGraph — Agent State Graph
# -------------------------------------------------------------
langgraph>=0.2.28

# -------------------------------------------------------------
# LLM Providers
# -------------------------------------------------------------
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
openai>=1.40.0,<2.0.0
anthropic>=0.40.0

# -------------------------------------------------------------
# Vector Store — Pinecone
# -------------------------------------------------------------
langchain-pinecone>=0.2.0
pinecone-client>=4.1.2

# -------------------------------------------------------------
# Document Loading (PDF)
# LlamaIndex SimpleDirectoryReader kept for PDF ingestion only
# -------------------------------------------------------------
llama-index-core>=0.12.0
llama-index-readers-file>=0.4.0
pypdf>=4.3.1
pdfminer.six>=20231228

# -------------------------------------------------------------
# Data Validation — Pydantic (used by router structured output)
# -------------------------------------------------------------
pydantic>=2.8.2

# -------------------------------------------------------------
# Environment & Configuration
# -------------------------------------------------------------
python-dotenv>=1.0.1

# -------------------------------------------------------------
# UI
# -------------------------------------------------------------
streamlit>=1.38.0

# -------------------------------------------------------------
# Evaluation & Testing
# -------------------------------------------------------------
pytest>=8.3.3
pytest-mock>=3.14.0

# -------------------------------------------------------------
# Utilities
# -------------------------------------------------------------
tqdm>=4.66.5
typing-extensions>=4.12.2